refactor and attention logger
