+++ src/rmind/callbacks/loggers/__init__.py
@@ -1,0 +1,15 @@
+from rmind.components.mask import WandbAttentionMaskLegend
+
+from .attention_mask import WandbAttentionMaskLogger, visualize_attention_mask
+from .image_param import WandbImageParamLogger
+from .patch_similarity import WandbPatchSimilarityLogger
+from .waypoints import WandbWaypointsLogger
+
+__all__ = [
+    "WandbAttentionMaskLegend",
+    "WandbAttentionMaskLogger",
+    "WandbImageParamLogger",
+    "WandbPatchSimilarityLogger",
+    "WandbWaypointsLogger",
+    "visualize_attention_mask",
+]
